requirements.txt
fastapi
uvicorn
SQLAlchemy
python-multipart
passlib[bcrypt]
python-jose
email-validator
pydantic
pytest
httpx
celery
redis
elasticsearch==8.10.1
